change detailTalent views

+++ app/src/main/java/com/dylan/friendease/data/model/TalentResponse.kt
@@ -17,7 +17,7 @@ data class TalentData(
     @SerializedName("verified_status")
     val verifiedStatus: String,
     @SerializedName("rating")
-    val rating: Int?,
+    val rating: Float?,
     @SerializedName("LocationTalent")
     val location: LocationTalent,
     @SerializedName("auth")
